vl-safe/requirements.txt
# 数据集适配器依赖包
pandas>=1.5.0
pyarrow>=10.0.0
google-generativeai>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiohttp>=3.8.0
openai>=1.0.0
requests>=2.28.0
opencv-python>=4.8.0
numpy>=1.24.0

